Moved CQRS RMs and entity manager to `Storage` NS
