+++ src/Storage/ServicesUpdateResourceModel.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace RebelCode\EddBookings\Services;
+namespace RebelCode\EddBookings\Services\Storage;
 
 use Dhii\Exception\CreateOutOfRangeExceptionCapableTrait;
 use Dhii\Expression\LiteralTermInterface;
